Rename gulpfile dir
